HTML, CSS und JS Parts aus Program.cs entfernt und ordentlich organisiert. Program.cs wurde entsprechend angepasst.
