+++ ServerAvailabilityTest/wwwroot/index.html
@@ -1,0 +1,19 @@
+﻿<!DOCTYPE html>
+<html>
+<head>
+    <title>Simple C# Web App</title>
+    <link rel="stylesheet" href="css/styles.css">
+</head>
+<body>
+    <div class='container'>
+        <h1>Your C# Web Application</h1>
+        <p>The application is running successfully!</p>
+
+        <button id='showMessageBtn'>Click Me!</button>
+
+        <div id='message'></div>
+
+        <script src="js/app.js"></script>
+    </div>
+</body>
+</html>
